fix: 修复vm-core的clippy警告
修复了vm-core包中的所有clippy编译错误：

1. pthread_qos_class_t命名问题
   - 添加#[allow(non_camel_case_types)]属性
   - 添加文档说明需要匹配Apple pthread API命名约定
   - 这些是FFI绑定，必须保持与系统API一致

2. GPU检测方法未使用警告
   - 为detect_cuda_device和detect_rocm_device添加#[allow(dead_code)]
   - 添加文档说明这些方法在特定feature下可用
   - 为所有feature配置提供实现（包括stub）

修复前: 7个clippy编译错误
修复后: 0个错误，仅剩可忽略的警告

相关文档: docs/CLIPPY_FIXES_SESSION_2026_01_06.md

+++ vm-engine-jit/src/parallel_compiler.rs
@@ -272,7 +272,7 @@ impl ParallelJITCompiler {
             total_blocks: stats.compiled_blocks,
             total_time_ns: stats.total_compile_time_ns,
             avg_block_size: if stats.compiled_blocks > 0 {
-                (stats.generated_code_size / stats.compiled_blocks as u64) as usize
+                (stats.generated_code_size / stats.compiled_blocks) as usize
             } else {
                 0
             },
